Added measurement
Writes a .csv file to the devices system to store the vaues of Input.acceleration or something like this. Needed to test, at which speed the vritual model doesn't stay at its place and follows the camera, to give the user some feedback if he is moving the device too fast.