Iteration3 so far so good
Added code to attach items to the componentes of the virtual model and manually fill the items. Additionally will the compononents information now be shown in the componenmenu, after it was clicked.
Changed VirtualModel_Toggle button to not set the virtualModel inactive, what would disable all scripts.
Components can now be selected and deselected and selected components can be completely disabled, so components behind this component can get clicked.