Added new scenes
Proof: Model Target which only overlays a cube (proofs that even a simple object follows the camera)
ModelTargetGroundPlaneScene (short MTGP): not working yet, would have a Model Target but the virtual model gets added to a ground plane after the position of the real model was found by the model target.
MainScene: the new Startupscene which allows the user to choose which scene/test he wants to open up
SampleScene: the previously done basic Model Target with the overlaying virtual model which is clickable, hidable and paintable.